created blank slate
