--- GitStats/embed.css
@@ -1,3 +1,0 @@
-* {
-    font-family: "Public Sans";
-}
