--- GitStats/embed_home.html
@@ -1,5 +1,0 @@
-<link href="font.css" rel="stylesheet">
-<link href="embed.css" rel="stylesheet">
-<p>Home page</p>
-<p id="loginresult">Verifying token...</p>
-<script src="./renderer_home.js"></script>
