--- GitStats/embed_issuemaker.html
@@ -1,16 +1,0 @@
-<link href="font.css" rel="stylesheet">
-<link href="embed.css" rel="stylesheet">
-<style>
-    #centered {
-        text-align: center;
-    }
-</style>
-<p id="centered" style="font-weight:700; font-size:30px; padding-bottom:30px">Issue Creator</p>
-<form onsubmit="return false" id="form">
-    <div id="options"></div>
-    <input type="text" placeholder="Title" id="issuetitle"><br>
-    <textarea id="issuedesc" cols="50" rows="10" placeholder="Description (supports markdown)" style="resize:none"></textarea><br>
-    <input type="submit" value="Create issue(s)" onclick="submitissues()">
-</form>
-<div id="successcontainer"></div>
-<script src="./renderer_issuemaker.js"></script>
