--- GitStats/embed_login.html
@@ -1,25 +1,0 @@
-<style>
-    #centered {
-        text-align: center;
-    }
-</style>
-<script>
-    var shell = window.electron.shell;
-    document.getElementById("howgettoken").setAttribute("onclick", `shell.openExternal("https://docs.github.com/en/enterprise-server@3.4/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token#about-personal-access-tokens")`)
-</script>
-<p id="centered" style="font-weight:700; font-size:30px; padding-bottom:30px">GitHub Account</p>
-<img src=".\assets\github.png" id="logo" width=100px style="margin-left: auto; margin-right: auto; display: block;">
-<p id="centered" style="font-size:12px">
-    <a href="#" id="howgettoken"><i style="font-size:12px">How do I get my GitHub token?</i><br></a>
-    <i>When generating a new token using the "Fine Grained" tokens, please give access to all repositories, and make sure to give Read & Write access to Issues.<br>This may be replaced with a single "give me my token" button in the future.</i>
-</p>
-<form id="centered" onsubmit="return false">
-    <input type="password" placeholder="GitHub Token" id="gittoken" name="token">
-    <input type="checkbox" onclick="showtoken()">Show token
-    <br><br> <!-- now that's what I call webdev -->
-    <input type="button" value="Login" id="loginbutton">
-    <br>
-    <input type="button" value="Reset my Local Data" onclick="resetuserdata()">
-    <p><i>careful! this button has no confirmation.</i></p>
-</form>
-<script src="./renderer_login.js"></script>
