--- GitStats/embed_stats.html
@@ -1,15 +1,0 @@
-<link href="font.css" rel="stylesheet">
-<link href="embed.css" rel="stylesheet">
-
-<form onsubmit="return false" id='reposelector'>
-    <p>Select a repository.</p>
-    <select id="repodropdown" onchange="addGraphOfSelectedRepo()">
-        
-    </select>
-    <!-- <input type="text" placeholder="Filters..." id='filters'> -->
-    <input type="submit" value="Get/Refresh Data" id="searchbutton" onclick="getData()">
-</form>
-<script src="./renderer_stats.js"></script>
-<div id="datacontainer">
-
-</div>
