--- GitStats/font.css
@@ -1,4 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Public+Sans&display=swap');
-* {
-    font-family: "Public Sans";
-}
