--- GitStats/index.css
@@ -1,25 +1,0 @@
-#topbar
-{
-    width:100%;
-    height: 50px;
-    background:rgb(22, 27, 34);
-    position: fixed;
-    left: 0px;
-    top: 0px;
-}
-#topbar .button
-{
-    text-align: center;
-    vertical-align: middle;
-    height:100%;
-    width: 24.61%; /*sloppy fix*/
-    display:inline-block;
-    color: white;
-    background-color: rgb(22, 27, 34);
-    border: 0px;
-    font-family: "Public Sans";
-    font-weight: 900;
-}
-#topbar .button:hover{
-    background-color: rgb(60, 70, 84);
-}
