--- GitStats/index.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="UTF-8">
-        <title>GitStats</title>
-        <meta http-equiv="Content-Security-Policy" content="script-src 'self' 'unsafe-inline';" />
-
-        <link rel="preconnect" href="https://fonts.googleapis.com">
-        <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-        <link rel="stylesheet" href="./font.css">
-        <link rel="stylesheet" href="./index.css">
-    </head>
-    <body>
-        <div id='topbar'>
-            <button class="button" load="embed_login.html">Account</button>
-            <button class="button" load="embed_issueload.html">Repositories</button>
-            <button class="button" load="embed_stats.html">Statistics</button>
-            <button class="button" load="embed_issuemaker.html">Issue Creator</button>
-        </div>
-        <div style="width:100%; height:50px">
-            &nbsp;
-        </div>
-        <div id="container">
-            uh
-        </div>
-    </body>
-    <script src="./renderer.js"></script>
-</html>
