+++ GitStats/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "GitStats.App",
+  "name": "gitstats",
   "productName": "GitStats",
-  "description": "An app written for my IA of my Senior High School Year 2023.",
+  "description": "A Lightweight ElectronJS app to get statistics about a repository.",
   "keywords": [],
-  "main": "./main.js",
-  "version": "1.0.0",
+  "main": "./main.ts",
+  "version": "2.0.0",
   "author": "megamaz",
   "scripts": {
     "start": "electron-forge start",
@@ -22,6 +22,7 @@
     "@electron-forge/maker-rpm": "^6.0.5",
     "@electron-forge/maker-squirrel": "^6.0.5",
     "@electron-forge/maker-zip": "^6.0.5",
-    "electron": "23.1.1"
+    "electron": "23.1.1",
+    "typescript": "^5.2.2"
   }
 }
