+++ GitStats/pages/index.html
@@ -1,0 +1,7 @@
+<!DOCTYPE html>
+
+<html>
+    <body>
+        <p>Beep Boop we're starting from scratch!</p>
+    </body>
+</html>
